Move FileSet to seperate file;
